Trustless KeyGen (#26)
In order to make it so the KeyGen process cannot be DoS'ed by a
malicious participant, we make it fully permissionless. In particular,
the group ID is now deterministically computed based on the group
parameters (with an added `context` for disambiguation). This allows the
epoch rotation KeyGen ceremony to happen without the consensus contract
at all, where the validators themselves are responsible for doing the
KeyGen ceremony and just proposing a rollover to the consensus contract.
In case there is a dishonest participant, the validators can
automatically kick off a new KeyGen ceremony with the dishonest
participant removed from the set.

In order to simplify the validators doing a key generation, a new
`keyGenAndCommit` method was added. It initializes the KeyGen storage
iff it has not already been initialized (instead of reverting like it
does for the `keyGen` function) and providing a DKG commitment. This
helps remove ambiguity from "who is responsible for starting KeyGen"
among the validators, and lets them all kickoff the process without
waiting for one another. The validators themselves are responsible for
determining what the `context` is - tentatively it can be
`address(Consensus) || epoch`.

+++ contracts/src/lib/FROSTNonceCommitmentSet.sol
@@ -15,8 +15,8 @@ library FROSTNonceCommitmentSet {
     }
 
     struct Commitments {
-        uint32 next;
-        mapping(uint32 chunk => Root) chunks;
+        uint64 next;
+        mapping(uint64 chunk => Root) chunks;
     }
 
     type Root is bytes32;
@@ -30,14 +30,14 @@ library FROSTNonceCommitmentSet {
     /// @notice Commits to the next chunk of nonces, given the current signature
     ///         sequence for a group. This prevents participants commiting to
     ///         nonces _after_ a signing ceremony has already begun.
-    function commit(T storage self, FROST.Identifier identifier, bytes32 commitment, uint32 sequence)
+    function commit(T storage self, FROST.Identifier identifier, bytes32 commitment, uint64 sequence)
         internal
-        returns (uint32 chunk)
+        returns (uint64 chunk)
     {
         Commitments storage commitments = self.commitments[identifier];
         uint256 offset;
         (chunk, offset) = _sequence(sequence);
-        uint32 next = commitments.next;
+        uint64 next = commitments.next;
         if (next > chunk) {
             chunk = next;
         }
@@ -51,13 +51,13 @@ library FROSTNonceCommitmentSet {
         FROST.Identifier identifier,
         Secp256k1.Point memory d,
         Secp256k1.Point memory e,
-        uint32 sequence,
+        uint64 sequence,
         bytes32[] calldata proof
     ) internal view {
         d.requireNonZero();
         e.requireNonZero();
 
-        (uint32 chunk, uint256 offset) = _sequence(sequence);
+        (uint64 chunk, uint256 offset) = _sequence(sequence);
         (bytes32 commitment, uint256 startOffset) = _root(self.commitments[identifier].chunks[chunk]);
         require(offset >= startOffset, NotIncluded());
 
@@ -80,7 +80,7 @@ library FROSTNonceCommitmentSet {
         }
     }
 
-    function _sequence(uint32 sequence) private pure returns (uint32 chunk, uint256 offset) {
+    function _sequence(uint64 sequence) private pure returns (uint64 chunk, uint256 offset) {
         chunk = sequence >> _CHUNKSZ;
         offset = uint256(sequence) & _OFFSETMASK;
     }
